requirements.txt
flask
flask-login
flask-bcrypt
flask-sqlalchemy
python-dotenv
gunicorn
requests
reportlab
